+++ src/components/common/join/Modal/JoinModal.tsx
@@ -7,6 +7,7 @@ import {
 } from '@hooks/query/useMutateQuery';
 import { useRecoilValue } from 'recoil';
 import { UserState } from '@store/user';
+import { useApplyStudyCase } from '@hooks/useApplyStudyCase';
 
 interface IJoinModalProps {
   isOpen: boolean;
@@ -16,23 +17,8 @@ interface IJoinModalProps {
 
 export function JoinModal({ isOpen, setIsOpen, postId }: IJoinModalProps) {
   const useProfile = useRecoilValue(UserState);
-  const [formData, setFormData] = useState('');
-  const { mutateAsync: addApplicantMutateAsync } = useUpdateApplicant(
-    Number(postId),
-  );
-  const applyApplicationQuery = useDeleteApplicant(postId);
-
-  const addApplicationOnClick = () => {
-    addApplicantMutateAsync(formData, {
-      onSuccess: () => {
-        setIsOpen((pre) => !pre);
-      },
-    });
-  };
-  const onChangeByTextarea = (e: ChangeEvent<HTMLTextAreaElement>) => {
-    const { value } = e.target;
-    setFormData(value);
-  };
+  const { formData, onChangeByTextarea, addApplicationOnClick } =
+    useApplyStudyCase(postId, setIsOpen);
 
   return (
     <Modal isOpen={isOpen} setIsOpen={setIsOpen}>
@@ -48,7 +34,7 @@ export function JoinModal({ isOpen, setIsOpen, postId }: IJoinModalProps) {
           <textarea
             maxLength={500}
             className="ApplyIntro"
-            placeholder="본인에 대한 짧은 소개입니다. 본인에 대한 짧은 소개입니다. 본인에 대한 짧은 소개입니다. 본인에 대한 짧은 소개입니다."
+            placeholder="본인에 대한 짧은 소개입니다."
             value={formData}
             onChange={onChangeByTextarea}
           />
